pyproject.toml
[project]
name = "uit-ds310"
version = "0.1.0"
description = "Vietnamese Legal Document Retrieval System"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "UIT-DS310"}
]
keywords = ["legal", "retrieval", "reranking", "vietnamese", "nlp"]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
vllm = [
    "vllm>=0.9.0",  # Chỉ cài nếu dùng QwenRerankerVLLM (rất nặng, cần compile)
]